requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
PyJWT
argon2-cffi
anthropic
voyageai
pgvector